requirements.txt
ta-lib==0.4.29
alpaca-py==0.33.0
pytest==8.2.1
pandas==2.2.2
PyYAML==6.0.1
pre_commit==4.0.1
loguru==0.7.2
